Move listen.py to the root (we need to refactor and move python stuff into a listenbrainz subdir) so that you don't need to install flask and other silly things just to use the Listen object.
