+++ listenstore/listenstore/tests/test_postgreslistenstore.py
@@ -4,7 +4,7 @@ from db.testing import DatabaseTestCase
 import logging
 from datetime import datetime
 from .util import generate_data, to_epoch
-from listenstore.listen import Listen
+from listen import Listen
 from listenstore.listenstore import PostgresListenStore, MIN_ID
 from webserver.postgres_connection import init_postgres_connection
 import random
